requirements.txt
streamlit==1.31.0
groq==0.4.2
python-dotenv==1.0.0
pandas==2.2.0
sqlalchemy==2.0.25
Pillow==10.2.0
plotly==5.18.0
